Tell gcc to avoid floating point contraction
On ppc64el Ubuntu, gcc was using fmadd instead of fmul which caused
'-459.67 °F' to be parsed as '-8.18283712394279 fK' instead of plain
'0 K'. The same problem with Reaumur could be fixed by adding extra
precision digits to the numbers in the temperature.sql file, but that
didn't work for °F. Turn add/mul contraction off for the nicer-looking
result.